+++ shuup/default_reports/reports/__init__.py
@@ -1,0 +1,12 @@
+# -*- coding: utf-8 -*-
+# This file is part of Shuup.
+#
+# Copyright (c) 2012-2016, Shoop Ltd. All rights reserved.
+#
+# This source code is licensed under the AGPLv3 license found in the
+# LICENSE file in the root directory of this source tree.
+from .sales import SalesReport
+from .sales_per_hour import SalesPerHour
+from .total_sales import TotalSales
+
+__all__ = ["SalesReport", "TotalSales", "SalesPerHour"]